feat: completed custom universe object & startup script
- update to Isaac Sim 4.2.0
- adios `World` object, welcome `Universe`!
- singular `big_bang` function that creates the universe singleton
- streamlined API for physics device query
- expanded configuration physics options
- streamlined terrain mesh insertion into USD stage

+++ core/animation/animation_engine.py
@@ -104,7 +104,7 @@ class AnimationEngine:
 
     def get_current_clip_data_ordered(
         self,
-        progress: int,
+        progress: float,
         joints_order: List[str],
         interpolate: bool = True,
     ) -> np.ndarray:
